Move SpecimenHeaderBean and SpecimenViewTypeForm to study-api. Prep for SpecimensAction move.
