+++ study/src/org/labkey/study/controllers/specimen/SpecimenController.java
@@ -28,7 +28,6 @@ import org.labkey.api.attachments.AttachmentFile;
 import org.labkey.api.attachments.ByteArrayAttachmentFile;
 import org.labkey.api.data.BaseColumnInfo;
 import org.labkey.api.data.ButtonBar;
-import org.labkey.api.data.ColumnInfo;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.DataRegionSelection;
 import org.labkey.api.data.ExcelWriter;
@@ -39,7 +38,6 @@ import org.labkey.api.data.TableInfo;
 import org.labkey.api.pipeline.browse.PipelinePathForm;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.query.QueryParam;
-import org.labkey.api.query.QueryService;
 import org.labkey.api.query.QueryUpdateForm;
 import org.labkey.api.query.QueryView;
 import org.labkey.api.query.UserSchemaAction;
@@ -56,6 +54,8 @@ import org.labkey.api.specimen.SpecimenRequestManager.SpecimenRequestInput;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.actions.IdForm;
 import org.labkey.api.specimen.actions.ParticipantCommentForm;
+import org.labkey.api.specimen.actions.SpecimenHeaderBean;
+import org.labkey.api.specimen.actions.SpecimenViewTypeForm;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
 import org.labkey.api.specimen.model.SpecimenRequestActor;
@@ -75,6 +75,7 @@ import org.labkey.api.specimen.settings.SettingsManager;
 import org.labkey.api.study.CohortFilter;
 import org.labkey.api.study.SpecimenUrls;
 import org.labkey.api.study.Study;
+import org.labkey.api.study.StudyInternalService;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.study.StudyUrls;
 import org.labkey.api.study.TimepointType;
@@ -122,13 +123,10 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Set;
 
 /**
@@ -192,8 +190,8 @@ public class SpecimenController extends BaseStudyController
         public ActionURL getCommentURL(Container c, String globalUniqueId)
         {
             return getSpecimensURL(c)
-                .addParameter(SpecimenController.SpecimenViewTypeForm.PARAMS.showVials, true)
-                .addParameter(SpecimenController.SpecimenViewTypeForm.PARAMS.viewMode, SpecimenQueryView.Mode.COMMENTS.name())
+                .addParameter(SpecimenViewTypeForm.PARAMS.showVials, true)
+                .addParameter(SpecimenViewTypeForm.PARAMS.viewMode, SpecimenQueryView.Mode.COMMENTS.name())
                 .addParameter("SpecimenDetail.GlobalUniqueId~eq", globalUniqueId);
         }
 
@@ -420,13 +418,14 @@ public class SpecimenController extends BaseStudyController
         @Override
         protected ModelAndView getHtmlView(SpecimenViewTypeForm form, BindException errors) throws Exception
         {
-            StudyImpl study = getStudyRedirectIfNull();
+            Study study = getStudyRedirectIfNull();
 
             SpecimenQueryView view = createInitializedQueryView(form, errors, form.getExportType() != null, null);
             SpecimenHeaderBean bean = new SpecimenHeaderBean(getViewContext(), view);
             // Get last selected request
-            if (null != study.getLastSpecimenRequest())
-                bean.setSelectedRequest(study.getLastSpecimenRequest());
+            Integer lastSpecimenRequest = StudyInternalService.get().getLastSpecimenRequest(study);
+            if (null != lastSpecimenRequest)
+                bean.setSelectedRequest(lastSpecimenRequest);
             JspView<SpecimenHeaderBean> header = new JspView<>("/org/labkey/study/view/specimen/specimenHeader.jsp", bean);
             return new VBox(header, view);
         }
@@ -436,167 +435,6 @@ public class SpecimenController extends BaseStudyController
         {
             addBaseSpecimenNavTrail(root);
             root.addChild(_vialView ? "Vials" : "Grouped Vials");
-        }
-    }
-
-    public static final class SpecimenHeaderBean
-    {
-        private final ActionURL _otherViewURL;
-        private final ViewContext _viewContext;
-        private final boolean _showingVials;
-        private final Set<Pair<String, String>> _filteredPtidVisits;
-
-        private Integer _selectedRequest;
-
-        public SpecimenHeaderBean(ViewContext context, SpecimenQueryView view)
-        {
-            this(context, view, Collections.emptySet());
-        }
-
-        public SpecimenHeaderBean(ViewContext context, SpecimenQueryView view, Set<Pair<String, String>> filteredPtidVisits) throws RuntimeException
-        {
-            Map<String, String[]> params = context.getRequest().getParameterMap();
-
-            String currentTable = view.isShowingVials() ? "SpecimenDetail" : "SpecimenSummary";
-            String otherTable =   view.isShowingVials() ? "SpecimenSummary" : "SpecimenDetail";
-            ActionURL otherView = context.cloneActionURL();
-            otherView.deleteParameters();
-
-            StudyImpl study = getStudy(context.getContainer());
-            if (null == study)
-                throw new NotFoundException("No study exists in this folder.");
-            StudyQuerySchema schema = StudyQuerySchema.createSchema(study, context.getUser(), true);
-
-            TableInfo otherTableInfo = schema.getTable(otherTable);
-
-            for (Map.Entry<String, String[]> param : params.entrySet())
-            {
-                int dotIndex = param.getKey().indexOf('.');
-
-                if (dotIndex >= 0)
-                {
-                    String table = param.getKey().substring(0, dotIndex);
-                    String columnClause = param.getKey().substring(dotIndex + 1);
-                    String[] columnClauseParts = columnClause.split("~");
-                    String column = columnClauseParts[0];
-
-                    if (table.equals(currentTable))
-                    {
-                        // use the query service to check to see if the current filter column is present
-                        // in the other view. If so, we'll add a filter parameter with the same value on the
-                        // other view. Otherwise, we'll keep the parameter, but we won't map it to the other view:
-                        boolean translatable = column.equals("sort");
-
-                        if (!translatable)
-                        {
-                            Map<FieldKey, ColumnInfo> presentCols = QueryService.get().getColumns(otherTableInfo,
-                                    Collections.singleton(FieldKey.fromString(column)));
-                            translatable = !presentCols.isEmpty();
-                        }
-
-                        if (translatable)
-                        {
-                            String key = otherTable + "." + columnClause;
-                            otherView.addParameter(key, param.getValue()[0]);
-                            continue;
-                        }
-                    }
-
-                    if (table.equals(currentTable) || table.equals(otherTable))
-                        otherView.addParameter(param.getKey(), param.getValue()[0]);
-                }
-            }
-
-            otherView.replaceParameter("showVials", Boolean.toString(!view.isShowingVials()));
-            if (null != params.get(SpecimenQueryView.PARAMS.excludeRequestedBySite.name()))
-                otherView.replaceParameter(SpecimenQueryView.PARAMS.excludeRequestedBySite.name(),
-                        params.get(SpecimenQueryView.PARAMS.excludeRequestedBySite.name())[0]);
-            _otherViewURL = otherView;
-            _viewContext = context;
-            _showingVials = view.isShowingVials();
-            _filteredPtidVisits = filteredPtidVisits;
-        }
-
-        public Integer getSelectedRequest()
-        {
-            return _selectedRequest;
-        }
-
-        public void setSelectedRequest(Integer selectedRequest)
-        {
-            _selectedRequest = selectedRequest;
-        }
-
-        public ActionURL getOtherViewURL()
-        {
-            return _otherViewURL;
-        }
-
-        public ViewContext getViewContext()
-        {
-            return _viewContext;
-        }
-
-        public boolean isShowingVials()
-        {
-            return _showingVials;
-        }
-
-        public Set<Pair<String, String>> getFilteredPtidVisits()
-        {
-            return _filteredPtidVisits;
-        }
-
-        public boolean isSingleVisitFilter()
-        {
-            if (getFilteredPtidVisits().isEmpty())
-                return false;
-            Iterator<Pair<String, String>> visitIt = getFilteredPtidVisits().iterator();
-            String firstVisit = visitIt.next().getValue();
-            while (visitIt.hasNext())
-            {
-                if (!Objects.equals(firstVisit, visitIt.next().getValue()))
-                    return false;
-            }
-            return true;
-        }
-    }
-
-    public static class SpecimenViewTypeForm extends QueryViewAction.QueryExportForm
-    {
-        public enum PARAMS
-        {
-            showVials,
-            viewMode
-        }
-
-        private boolean _showVials;
-        private SpecimenQueryView.Mode _viewMode = SpecimenQueryView.Mode.DEFAULT;
-
-        public boolean isShowVials()
-        {
-            return _showVials;
-        }
-
-        public void setShowVials(boolean showVials)
-        {
-            _showVials = showVials;
-        }
-
-        public String getViewMode()
-        {
-            return _viewMode.name();
-        }
-
-        public SpecimenQueryView.Mode getViewModeEnum()
-        {
-            return _viewMode;
-        }
-
-        public void setViewMode(String viewMode)
-        {
-            if (viewMode != null)
-                _viewMode = SpecimenQueryView.Mode.valueOf(viewMode);
         }
     }
